fix: sysroot path should be the target compiler
We use the `host` compiler (which is going to be `x86_64`), while
linking against the `target`runtime (which is `aarch64`) to produce the
Lean toolchain.

The `sysroot` option sets the path for the `host` compiler to pickup the
runtimes for the `target`? So in our case, the `sysroot` should point to
the `aarch64` runtime (ie, the `target` runtime).