fix: think through how llvm-config should be used.
How did this stuff ever work on linux?
I need to run `llvm-target/bin/llvm-config` to find out the linker
options I need for the target LLVM? But the target `llvm-config`
is an `aarch64` binary, which we definitely can't run on the host
`x86_64` machine!

Do we take on faith that `llvm-host/bin/llvm-config` will provide
flags that are approximately correct, and move ahead with that?
That seems like a dubious assumption to me.

It seems like it used to work on linux, as I had missed the
`-DLLVM_CONFIG=path/to/llvm-config` option in the linux builds.
So it was (luckily) finding the (host?) `llvm-config` and configuring
itself correctly?