fix: add " -L ROOT/llvm/lib" to Leanc linker flags
We see that on the build runner, we get the flags:

+ lean-4.0.0-linux/bin/leanc --print-ldflags
-I /home/runner/work/lean4/lean4/lean-4.0.0-linux/include -fPIC -fvisibility=hidden -L /home/runner/work/lean4/lean4/lean-4.0.0-linux/lib/lean -Wl,--start-group -lleancpp -lLean -Wl,--end-group -Wl,--start-group -lInit -lleanrt -Wl,--end-group -Wl,-Bstatic -lc++ -lc++abi -Wl,-Bdynamic -L/home/runner/work/lean4/lean4/build/llvm/lib -lLLVM-15 -lm -Wl,--as-needed -lgmp -Wl,--no-as-needed -ldl -pthread

See that we have -L:
1. /home/runner/work/lean4/lean4/build/llvm/lib
2. /home/runner/work/lean4/lean4/lean-4.0.0-linux/lib/lean

The path (1) is generated by
`llvm-config`, relative to its location. It's also completely bunk.
If we want to generate
a sensible path from `llvm-config`, we should keep `llvm-config` within the `stage1/bin`
tree, and then query `stage1/bin/llvm-config --libs` or whatever.

Alternatively, We add another path like (2), expect to point to:
/home/runner/work/lean4/lean4/lean-4.0.0-linux/lib/llvm/

which will give it access to the LLVM libs.

This commit attempts to perform solution (2), which adds a new path.
I am not a CMake expert, so this might take a couple tries to get right.