fix: add libleanrt to shell build
We need `libleanrt` to provide us a runtime when emitting code.
This allows us to correctly build the runtime, and error out politely,
at the location of the mismatch, if we can't access clang as the C++
compiler.

There used to be a global requirement that the LLVM backend needed
clang, which I removed as legacy. I now see why it was potentially
required (to build `libleanrt.bc`). Hopefully, placing the check
for Clang and the erroring out here makes it clearer to the next person
as to why we enforce the need for Clang when building the LLVM backend.