fix: generate lean_inlines programatically
Using `#define static extern` caused lots of crazy
stuff to break in the header files. Instead, write a script
that copies `src/include/lean.h`, and then uses `sed` to replace
`static inline` in a controlled fashion. This ensures that the resulting
`.ll` file has all the definitions we want.