fix: codegen `initUnboxed` correctly in LLVM backend.
Closes #2004.

In porting the bugfix from
https://github.com/leanprover/lean4/commit/6eb852e28f2f224795d9a98fed8c84de0ebb4ca2,
I noticed that the LLVM backend was incorrectly generating declaration
initializers (in `callIODeclInitFn`), by assuming the return type of the
initializer is the return type of the declaration. Rather, it must be be
`lean_object`, since the initializer returns an `IO a` value which must be unpacked.

TODO: stop using the `getOrCreateFunction` pattern pervasively.
  perform the `create` at the right location, and the `get`
  at the correct location.