feat: lake: Sort manifest packages by name
This fixes #2664

The manifest’s `packages` field is only assembled in
`Workspace.updateAndMaterialize`, where previously a `OrdNameMap` was
filled. The `OrdNameMap` builds on `RBArray` which explicity preserves
the insertion order. But for the manifest, it seems desirable to keep
the entries independent of insertion order in this code, so this now
uses a name map that keeps names sorted.

Note that Lean’s `Lean.NameMap` is not a great choice, beceause it sorts
names by their Hash first; using this order would be counter-intuitive
for users, and possibly annoying if the hash function changes.

Therfore, we use `RBMap Name α Name.ord` (abbreviated to
`SortedNameMap`) here.